// src/domain/types/v1api/postListRankingSettingsBody.ts
/**
 * Generated by orval. Do not edit manually.
 * O-replay - OpenAPI 3.0
 * O-replay Rest API
 */

/**
 * Generic object when: Add new generates uuid server side
 */
export type PostListRankingSettingsBody = {
  event_id?: string
  excluded_class_names?: string
  max_points?: number
  nc_false?: number
  nc_true?: number
  overall_settings?: string
  round_precision?: number
  scoring_algorithm?: string
  stage_id?: string
  status_scores?: string
  title?: string
}